feat: recreate a home component
has all features that fit our usecase
